Move tests to new folder
