+++ src/components/tests/Dragons.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import Profile from './Profile';
-import store from '../redux/configureStore';
+import Dragons from '../Dragons';
+import store from '../../redux/configureStore';
 
-describe('Profile', () => { });
-it('Testing Profile component rendering', () => {
+describe('Dragons', () => { });
+it('Testing Dragons component rendering', () => {
   const component = render(
     <Provider store={store}>
-      <Profile />
+      <Dragons />
     </Provider>,
 
   );
