+++ src/components/tests/Header.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import Header from './Header';
+import Header from '../Header';
 
 describe('Header', () => { });
 it('Testing header component rendering', () => {
