+++ src/components/tests/Missions.test.js
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
-import store from '../redux/configureStore';
-import Mission from './Missions';
+import store from '../../redux/configureStore';
+import Mission from '../Missions';
 
 test('First Snapshot test', () => {
   const component = renderer
