+++ src/components/tests/Profile.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import Rockets from './Rockets';
-import store from '../redux/configureStore';
+import Profile from '../Profile';
+import store from '../../redux/configureStore';
 
-describe('Rockets', () => { });
-it('Testing Rockets component rendering', () => {
+describe('Profile', () => { });
+it('Testing Profile component rendering', () => {
   const component = render(
     <Provider store={store}>
-      <Rockets />
+      <Profile />
     </Provider>,
 
   );
